cloudfuncs/human_detection/requirements.txt
SQLAlchemy==1.3.13
psycopg2==2.8.4
google-cloud-storage==1.26.0
google-cloud-vision==0.42.0
requests==2.23.0